Fusion of nmpc visualization and improved differentiation between controller configs
